requirements.txt
Django==2.2.6
djangorestframework==3.10.3
django-rest-swagger
djangorestframework_simplejwt
django-oauth-toolkit
requests
django-filter